crud de endereços ok, crud de user a caminho

+++ src/app/endereco/index/index.component.ts
@@ -1,4 +1,7 @@
 import { Component, OnInit } from '@angular/core';
+import { ToastrService } from 'ngx-toastr';
+import { AlertService } from 'src/app/alert.service';
+import Swal from 'sweetalert2';
 import { Endereco } from '../endereco';
 import { EnderecoService } from '../endereco.service';
 
@@ -11,19 +14,37 @@ export class IndexComponent implements OnInit {
 
   enderecos: Endereco[] = [];
 
-  constructor(private enderecoService: EnderecoService) { }
+  constructor(
+    private enderecoService: EnderecoService,
+    private serviceAlert: AlertService
+  ) { }
 
   ngOnInit(): void {
+    this.getList();
+  }
+
+  getList() {
     this.enderecoService.getAll().subscribe((data) => {
       this.enderecos = data
+    }, (err) => {
+      this.serviceAlert.error('Listagem de endereços não pode ser carregada');
     });
   }
 
-
   delete(idEndereco: number) {
-    this.enderecoService.delete(idEndereco).subscribe(res => {
-      this.enderecos = this.enderecos.filter(item => item.id !== idEndereco);
+
+    this.serviceAlert.confirm().then((result) => {
+      if (result.isConfirmed) {
+        this.enderecoService.delete(idEndereco).subscribe(res => {
+          this.serviceAlert.success('Registro deletado com sucesso', 'Tudo Certo')
+          this.enderecos = this.enderecos.filter(item => item.id !== idEndereco);
+        }, (err) => {
+          this.serviceAlert.error('Não foi deletar', 'Erro');
+        });
+      }
     })
+
   }
+
 
 }
